init-db/requirements.txt
psycopg2-binary==2.7.5
scipy==1.1.0
